完善辅助文件清除函数

+++ src/pytexmk/additional_operation.py
@@ -16,7 +16,7 @@
  -----------------------------------------------------------------------
 Author       : 焱铭
 Date         : 2024-02-29 16:02:37 +0800
-LastEditTime : 2024-03-01 17:08:11 +0800
+LastEditTime : 2024-03-01 18:02:23 +0800
 Github       : https://github.com/YanMing-lxb/
 FilePath     : /PyTeXMK/src/pytexmk/additional_operation.py
 Description  : 
@@ -24,34 +24,43 @@ Description  :
 '''
 
 import os
-import datetime 
+import datetime
 import shutil
+
 # --------------------------------------------------------------------------------
 # 定义清除辅助文件命令
 # --------------------------------------------------------------------------------
 def remove_aux(file_name):
     auxiliary_files = [
-        f"{file_name}{ext}" for ext in [".pdf", ".synctex", ".aux", ".bbl", ".blg", ".log", ".out", ".toc", ".bcf",
-                                        ".xml", ".synctex", ".nlo", ".nls", ".bak", ".ind", ".idx", ".ilg", ".lof",
+        f"{file_name}{ext}" for ext in [".aux", ".bbl", ".blg", ".log", ".out", ".toc", ".bcf",
+                                        ".xml", ".nlo", ".nls", ".bak", ".ind", ".idx", ".ilg", ".lof",
                                         ".lot", ".ent-x", ".tmp", ".ltx", ".los", ".lol", ".loc", ".listing", ".gz",
                                         ".userbak", ".nav", ".snm", ".vrb", ".fls", ".xdv", ".fdb_latexmk", ".run.xml", ".ist", ".glo", ".gls"]
     ]
+    file_exists = False
     for file in auxiliary_files:
-        try:
-            os.remove(file)
-        except FileNotFoundError:
-            pass
-    print("已清除辅助文件")
+        if os.path.exists(file):
+            try:
+                os.remove(file)
+                file_exists = True
+            except FileNotFoundError:
+                pass
+    if file_exists:
+        print("已清除辅助文件")
+    else:
+        print("未找到辅助文件")
 
 # --------------------------------------------------------------------------------
 # 定义清除已有结果文件
 # --------------------------------------------------------------------------------
+
+
 def remove_result(build_path):
     if os.path.exists(build_path):
-        shutil.rmtree(build_path) # 删除整个文件夹
+        shutil.rmtree(build_path)  # 删除整个文件夹
         print("删除上次生成的结果文件")
-    os.mkdir(build_path) # 创建空的 build_path 文件夹
-    
+    os.mkdir(build_path)  # 创建空的 build_path 文件夹
+
 # --------------------------------------------------------------------------------
 # 定义移动生成文件
 # --------------------------------------------------------------------------------
@@ -80,11 +89,12 @@ def time_count(fun, *args):
 # 定义输入检查函数
 # --------------------------------------------------------------------------------
 def check_file_name(file_name):
-    base_name, file_extension = os.path.splitext(os.path.basename(file_name)) # 去掉路径，提取文件名和后缀
-    if file_extension == '.tex': # 判断后缀是否是 .tex
+    base_name, file_extension = os.path.splitext(
+        os.path.basename(file_name))  # 去掉路径，提取文件名和后缀
+    if file_extension == '.tex':  # 判断后缀是否是 .tex
         file_name_return = base_name
-    elif '.' not in file_name: # 判断输入 file_name 中没有 后缀
-        if '/' in file_name or '\\' in file_name: # 判断是否是没有后缀的路径
+    elif '.' not in file_name:  # 判断输入 file_name 中没有 后缀
+        if '/' in file_name or '\\' in file_name:  # 判断是否是没有后缀的路径
             file_name_return = None
             print("错误：文件路径无效")
         else:
@@ -92,7 +102,7 @@ def check_file_name(file_name):
     else:
         file_name_return = None
         print("提示：文件后缀不是.tex")
-    
+
     return file_name_return
 
 # --------------------------------------------------------------------------------
@@ -102,8 +112,9 @@ def search_file():
     # 获取当前路径
     current_path = os.getcwd()
     # 遍历当前路径下的所有文件
-    tex_files = [file for file in os.listdir(current_path) if file.endswith('.tex')]
-    
+    tex_files = [file for file in os.listdir(
+        current_path) if file.endswith('.tex')]
+
     if tex_files:
         # 如果只有一个.tex文件，则直接提取文件名并打印
         if len(tex_files) == 1:
@@ -123,5 +134,5 @@ def search_file():
         # 不存在.tex文件，打印当前路径并提示
         tex_name = None
         print("在当前终端路径下不存在 .tex 文件，请检查终端显示路径是否是项目路径:", current_path)
-    
+
     return tex_name
